Support related link groups; Improved association handling
git-svn-id: svn://kiwinb/dev/media/trunk@365 d2417a56-bda1-9d42-b968-744157609d61

+++ core/src/com/kiwisoft/media/photos/Photo.java
@@ -77,7 +77,7 @@ public class Photo extends IDObject implements ChainLink
 
 	public Long getOriginalPictureId()
 	{
-		return getReferenceId(ORIGINAL_PICTURE);
+		return (Long)getReferenceId(ORIGINAL_PICTURE);
 	}
 
 	public PictureFile getOriginalPicture()
